Konsultabot2-main/backend/requirements.txt
# Django Backend Requirements for KonsultaBot Advanced AI Platform
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-dotenv==1.0.0

# JWT Authentication
djangorestframework-simplejwt==5.3.1
PyJWT==2.8.0

# AI and NLP
google-generativeai>=0.3.0
nltk==3.8.1
sentence-transformers==2.2.2
langdetect==1.0.9
scikit-learn==1.3.2

# Optional dependencies (same as original project)
spacy==3.7.2
transformers==4.35.2
torch==2.1.1

# Database
psycopg2-binary==2.9.9  # For PostgreSQL (optional)

# Utilities
requests==2.31.0
Pillow==10.1.0
bcrypt==4.1.1
urllib3==2.0.7
